+++ test/ogf_comp_power_proxy.jl
@@ -1,9 +1,9 @@
 @testset "test new ogf" begin
     @testset "test wp new ogf" begin
         @testset "case 6 new ogf" begin
-            @info "Testing NEW OGF"
+            @info "Testing OGF with compressor power proxy (linear)"
             data = GasModels.parse_file("../test/data/matgas/case-6-no-power-limits.m")
-            result = run_new_ogf(data, WPGasModel, nlp_solver)
+            result = run_ogf_comp_power_proxy(data, WPGasModel, nlp_solver)
             @test result["termination_status"] in [LOCALLY_SOLVED, ALMOST_LOCALLY_SOLVED, OPTIMAL, :Suboptimal]
             @test isapprox(result["objective"], -253.683; atol = 1e-2)
             GasModels.make_si_units!(result["solution"])
@@ -11,9 +11,9 @@
         end
 
         @testset "case 6 new ogf weymouth lin rel" begin
-            @info "Testing NEW OGF Linear Relaxation of Pipe Weymouth Physics"
+            @info "Testing  OGF with compressor power proxy Linear Relaxation of Pipe Weymouth Physics"
             data = GasModels.parse_file("../test/data/matgas/case-6-no-power-limits.m")
-            result = run_new_ogf(data, LRWPGasModel, lp_solver)
+            result = run_ogf_comp_power_proxy(data, LRWPGasModel, lp_solver)
             @test result["termination_status"] in [LOCALLY_SOLVED, ALMOST_LOCALLY_SOLVED, OPTIMAL, :Suboptimal]
             @test isapprox(result["objective"], -260.00; atol = 1e-2)
             GasModels.make_si_units!(result["solution"])
@@ -22,9 +22,9 @@
 
 
         @testset "case 6 wp new ogf binding energy constraint" begin
-            @info "Testing NEW OGF Binding Energy Cosntraint"
+            @info "Testing  OGF with compressor power proxy Binding Energy Cosntraint"
             data = GasModels.parse_file("../test/data/matgas/case-6.m")
-            result = run_new_ogf(data, WPGasModel, nlp_solver)
+            result = run_ogf_comp_power_proxy(data, WPGasModel, nlp_solver)
             @test result["termination_status"] in [LOCALLY_SOLVED, ALMOST_LOCALLY_SOLVED, OPTIMAL, :Suboptimal]
             @test isapprox(result["objective"], -237.718; atol = 1e-2)
             GasModels.make_si_units!(result["solution"])
@@ -32,9 +32,9 @@
         end
 
         @testset "case 6 wp new ogf elevation constraint" begin
-            @info "Testing NEW OGF Elevation Cosntraint"
+            @info "Testing  OGF with compressor power proxy Elevation Cosntraint"
             data = GasModels.parse_file("../test/data/matgas/case-6-elevation.m")
-            result = run_new_ogf(data, WPGasModel, nlp_solver)
+            result = run_ogf_comp_power_proxy(data, WPGasModel, nlp_solver)
             @test result["termination_status"] in [LOCALLY_SOLVED, ALMOST_LOCALLY_SOLVED, OPTIMAL, :Suboptimal]
             @test isapprox(result["objective"], -245.251; atol = 1e-2)
             GasModels.make_si_units!(result["solution"])